src: remove req_wrap-inl.h from stream_base.h
This commit removes the inclusion of req_wrap-inl.h in stream_base.h
as ReqWrap is not used. This removal required stream_base.h to include
async_wrap-inl.h so there is an implementation of BaseObject::object.

The above change also affected connect_wrap, which needs to include
req_wrap-inl.h to get an implementation of ReqWrap::Dispatched.

PR-URL: https://github.com/nodejs/node/pull/20063
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Tiancheng "Timothy" Gu <timothygu99@gmail.com>