test: remove message argument from strictEqual()
In test/parallel/test-require-process.js, the last thing in the test is
a call to assert.strictEqual(). It has a string literal as its third
argument. Unfortunately, that means that the diff between the two values
being compared will be suppressed if there is an AssertionError. That's
not helpful for debugging.

This is fixed by removing the third argument from the call. It is,
however, preserved in a comment above the call to assert.strictEqual().

PR-URL: https://github.com/nodejs/node/pull/20912
Fixes: https://github.com/nodejs/node/issues/20911
Refs: https://www.nodetodo.org/getting-started
Reviewed-By: Benjamin Gruenbaum <benjamingr@gmail.com>
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: Ujjwal Sharma <usharma1998@gmail.com>
Reviewed-By: Ruben Bridgewater <ruben@bridgewater.de>