test: update test-dns error message
test-dns is in the internet suite and therefore is rarely run. (That
will change soon. We will run internet tests nightly in CI.) Because it
is infrequently run, it was not noticed that it no longer passes. (An
error message has changed to provide more information.) This change
fixes the test so it passes.

PR-URL: https://github.com/nodejs/node/pull/21116
Reviewed-By: Anatoli Papirovski <apapirovski@mac.com>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>